refatoramos a página de resultados
