+++ src/pages/HomePage/HomePage.js
@@ -9,7 +9,7 @@ import {
   Grid,
   GridItem,
   Heading,
-  IconButton,
+  Image,
   Input,
   InputGroup,
   InputLeftElement,
@@ -22,8 +22,6 @@ import {
   Text,
 } from "@chakra-ui/react";
 import { Search2Icon } from "@chakra-ui/icons";
-import { AiOutlineHome, AiOutlineShoppingCart } from "react-icons/ai";
-import { MdPersonOutline } from "react-icons/md";
 import {
   goToCartPage,
   goToHome,
@@ -34,6 +32,9 @@ import { useNavigate } from "react-router-dom";
 import useForm from "../../hooks/useForm";
 import { auth } from "../../constants/auth";
 import { useProtectedPage } from "../../hooks/useProtectedPage";
+import selectedHomePageImage from "../../assets/selected-homepage.svg"
+import shoppingCartImage from "../../assets/shopping-cart.svg"
+import avatarImage from "../../assets/avatar.svg"
 
 export default function HomePage() {
   useProtectedPage();
@@ -85,20 +86,7 @@ export default function HomePage() {
   const onClickCard = (id) => {
     goToResultPage(navigate, id);
   };
-  // const renderedRestaurants = listRestaurants.map((restaurant) => {
-  //   return (
-  //     <RestaurantCard
-  //       key={restaurant.id}
-  //       logoUrl={restaurant.logoUrl}
-  //       description={restaurant.description}
-  //       deliveryTime={restaurant.deliveryTime}
-  //       shipping={restaurant.shipping}
-  //       name={restaurant.name}
-  //       category={restaurant.category}
-  //       onClick={() => onClickCard(restaurant.id)}
-  //     ></RestaurantCard>
-  //   );
-  // });
+
   const renderedArabicRestaurants = listRestaurants.map((restaurant) => {
     if (restaurant.category === "Árabe") {
       return (
@@ -245,82 +233,90 @@ export default function HomePage() {
   });
 
   return (
-    <Box maxH="100vh">
-      <Center>
-        <Heading
-          as="h3"
-          size={"16px"}
-          borderBottom="1px"
-          borderColor="gray.200"
-          w="328px"
-          h="44px"
-          textAlign={"center"}
-          marginBottom="8px"
-          mt="13px"
-        >
-          Rappi4
-        </Heading>
-      </Center>
-      <Center>
-        <InputGroup w="328px" h="56px">
-          <InputLeftElement
-            pointerEvents="none"
-            children={<Search2Icon color="gray.300" />}
-          />
-          <Input
-            type="search"
-            placeholder="Restaurante"
-            onChange={onChange}
-            value={form.searchRestaurant}
-            name={"searchRestaurant"}
-          />
-        </InputGroup>
-      </Center>
-      <Center h={"80vh"}>
-        <Tabs>
+    <Grid templateRows="0fr 0fr 2fr 1fr" maxH="100vh" h="100vh">
+      <GridItem h={'44px'}>
+        <Center>
+          <Heading
+            as="h3"
+            size={"16px"}
+            boxShadow={'0 0.5px 0 0 rgba(0, 0, 0, 0.25);'}
+            w={'100vw'}
+            h="44px"
+            textAlign={"center"}
+            marginBottom="8px"
+            mt="13px"
+          >
+            Rappi4
+          </Heading>
+        </Center>
+      </GridItem>
+      <GridItem>
+        <Center>
+          <InputGroup w="328px" h="56px" mt={'22px'}>
+            <InputLeftElement
+              pointerEvents="none"
+              children={<Search2Icon color="gray.300" />}
+            />
+            <Input
+              type="search"
+              placeholder="Restaurante"
+              onChange={onChange}
+              value={form.searchRestaurant}
+              name={"searchRestaurant"}
+              borderRadius={'2px'}
+            />
+          </InputGroup>
+        </Center>
+      </GridItem>
+      <Box overflowY={'scroll'} mb={'8px'}>
+        <GridItem>
           <Center>
-            <TabList overflowX={"scroll"} overflowY={"hidden"} w="100vw">
-              <Tab _selected={{ color: "#e86e5a" }}>Todos os Restaurantes</Tab>
-              <Tab _selected={{ color: "#e86e5a" }}>Árabe</Tab>
-              <Tab _selected={{ color: "#e86e5a" }}>Asiática</Tab>
-              <Tab _selected={{ color: "#e86e5a" }}>Hamburguer</Tab>
-              <Tab _selected={{ color: "#e86e5a" }}>Italiana</Tab>
-              <Tab _selected={{ color: "#e86e5a" }}>Sorvetes</Tab>
-              <Tab _selected={{ color: "#e86e5a" }}>Carnes</Tab>
-              <Tab _selected={{ color: "#e86e5a" }}>Baiana</Tab>
-              <Tab _selected={{ color: "#e86e5a" }}>Petiscos</Tab>
-              <Tab _selected={{ color: "#e86e5a" }}>Mexicana</Tab>
-            </TabList>
-          </Center>
+            <Tabs>
+              <Center>
+                <TabList overflowX={"scroll"} overflowY={"hidden"} w="100vw">
+                  <Tab _selected={{ color: "#e86e5a" }}>Todos os Restaurantes</Tab>
+                  <Tab _selected={{ color: "#e86e5a" }}>Árabe</Tab>
+                  <Tab _selected={{ color: "#e86e5a" }}>Asiática</Tab>
+                  <Tab _selected={{ color: "#e86e5a" }}>Hamburguer</Tab>
+                  <Tab _selected={{ color: "#e86e5a" }}>Italiana</Tab>
+                  <Tab _selected={{ color: "#e86e5a" }}>Sorvetes</Tab>
+                  <Tab _selected={{ color: "#e86e5a" }}>Carnes</Tab>
+                  <Tab _selected={{ color: "#e86e5a" }}>Baiana</Tab>
+                  <Tab _selected={{ color: "#e86e5a" }}>Petiscos</Tab>
+                  <Tab _selected={{ color: "#e86e5a" }}>Mexicana</Tab>
+                </TabList>
+              </Center>
 
-          {load ? (
-            <Center>
-              <Spinner size="xl" marginTop="20px" />
-            </Center>
-          ) : (
-            <TabPanels overflowY={"scroll"} h="60vh">
-              <TabPanel>
-                {filteredRestaurantsLists.length ? (
-                  filteredRestaurantsLists
-                ) : (
-                  <Center>
-                    <Text marginTop={"20px"}>Não Encontramos :(</Text>
-                  </Center>
-                )}
-              </TabPanel>
-              <TabPanel>{renderedArabicRestaurants}</TabPanel>
-              <TabPanel>{renderedAsianRestaurants}</TabPanel>
-              <TabPanel>{renderedBurgerRestaurants}</TabPanel>
-              <TabPanel>{renderedItalianRestaurants}</TabPanel>
-              <TabPanel>{renderedIceCreamRestaurants}</TabPanel>
-              <TabPanel>{renderedMeatRestaurants}</TabPanel>
-              <TabPanel>{renderedBaianoRestaurants}</TabPanel>
-              <TabPanel>{renderedSnackRestaurants}</TabPanel>
-              <TabPanel>{renderedMexicanRestaurants}</TabPanel>
-            </TabPanels>
-          )}
-        </Tabs>
-      </Center>
+              {load ? (
+                <Center>
+                  <Spinner size="xl" marginTop="20px" />
+                </Center>
+              ) : (
+                <TabPanels overflowY={"scroll"} h="60vh">
+                  <TabPanel>
+                    {filteredRestaurantsLists.length ? (
+                      filteredRestaurantsLists
+                    ) : (
+                      <Center>
+                        <Text marginTop={"20px"}>Não Encontramos :(</Text>
+                      </Center>
+                    )}
+                  </TabPanel>
+                  <TabPanel>{renderedArabicRestaurants}</TabPanel>
+                  <TabPanel>{renderedAsianRestaurants}</TabPanel>
+                  <TabPanel>{renderedBurgerRestaurants}</TabPanel>
+                  <TabPanel>{renderedItalianRestaurants}</TabPanel>
+                  <TabPanel>{renderedIceCreamRestaurants}</TabPanel>
+                  <TabPanel>{renderedMeatRestaurants}</TabPanel>
+                  <TabPanel>{renderedBaianoRestaurants}</TabPanel>
+                  <TabPanel>{renderedSnackRestaurants}</TabPanel>
+                  <TabPanel>{renderedMexicanRestaurants}</TabPanel>
+                </TabPanels>
+              )}
+            </Tabs>
+          </Center>
+        </GridItem>
+      </Box>
       <Center>
         <Grid
           templateColumns={"repeat(3, 1fr)"}
@@ -328,38 +324,44 @@ export default function HomePage() {
           borderWidth="1px"
           w={"100vw"}
           align="center"
+          justifyItems={'center'}
           position={"fixed"}
+          bottom={'27%'}
+          h={'49px'}
+          bg={'#fff'}
         >
-          <GridItem>
-            <IconButton
-              as={AiOutlineHome}
-              bg={"white"}
-              w={"27px"}
-              h={"27px"}
-              color={"#e865a"}
+          <GridItem
+            alignSelf={'center'}
+          >
+            <Image
+              src={selectedHomePageImage}
+              w={'27px'}
+              h={'27px'}
               onClick={() => goToHome(navigate)}
             />
           </GridItem>
-          <GridItem>
-            <IconButton
-              as={AiOutlineShoppingCart}
-              bg={"white"}
-              w={"27px"}
-              h={"27px"}
+          <GridItem
+            alignSelf={'center'}
+          >
+            <Image
+              src={shoppingCartImage}
+              w={'27px'}
+              h={'29px'}
               onClick={() => goToCartPage(navigate)}
             />
           </GridItem>
-          <GridItem>
-            <IconButton
-              as={MdPersonOutline}
-              bg={"white"}
-              w={"27px"}
-              h={"27px"}
+          <GridItem
+            alignSelf={'center'}
+          >
+            <Image
+              src={avatarImage}
+              w={'27px'}
+              h={'30px'}
               onClick={() => goToProfilePage(navigate)}
             />
           </GridItem>
         </Grid>
       </Center>
-    </Box>
+    </Grid>
   );
 }
