+++ src/pages/ResultPage/ResultPage.js
@@ -6,11 +6,8 @@ import { auth } from "../../constants/auth";
 import RestaurantDetailsCard from "../../components/RestaurantsCard/RestaurantsDetailsCard";
 import {
   Box,
-  Button,
-  Center,
   Container,
   Divider,
-  Flex,
   Grid,
   GridItem,
   Image,
@@ -28,6 +25,8 @@ import {
 import back from "../../assets/back.svg";
 import { goToHome } from "../../Routes/Coordinator";
 import { useProtectedPage } from "../../hooks/useProtectedPage";
+import ProductsCard from "../../components/ProductsCard/ProductsCard";
+import ModalAddButton from "../../components/ModalAddButton/ModalAddButton";
 
 export default function ResultPage() {
   useProtectedPage();
@@ -43,7 +42,6 @@ export default function ResultPage() {
     axios
       .get(`${BASE_URL}/restaurants/${id}`, auth)
       .then((res) => {
-        console.log(res.data.restaurant);
         setRestaurantDetails(res.data.restaurant);
         setProducts(res.data.restaurant.products);
       })
@@ -56,153 +54,22 @@ export default function ResultPage() {
     getRestaurantDetails(id);
   }, []);
 
-  const currency = function (number) {
-    return new Intl.NumberFormat("pt-BR", {
-      style: "currency",
-      currency: "BRL",
-      minimumFractionDigits: 2,
-    }).format(number);
-    //adriano rei do pt-br
-  };
-
-  const handleChange = (event) => {
-    setQuantity(event.target.value)
-  }
-
   const mainProductsMapped = products.map((product) => {
     if (
       product.category !== "Bebida" &&
       product.category !== "Acompanhamento"
     ) {
       return (
-        <Center>
-          <Flex
-            border="solid 1px #b8b8b8"
-            borderRadius="8px"
-            m="16px"
-            w="328px"
-            h="112px"
-            key={product.id}
-          >
-            <Image
-              w="97px"
-              maxW="97px"
-              h="112.6px"
-              maxH={"112.6px"}
-              borderTopLeftRadius="8px"
-              borderBottomLeftRadius="8px"
-              src={product.photoUrl}
-              alt={product.name}
-            />
-            <Box ml="3">
-              <Text
-                fontWeight="bold"
-                fontSize="16px"
-                color="#e86e5a"
-                margin={"18px 16px 6px 0"}
-                maxW={"166px"}
-                isTruncated
-                textOverflow={"ellipsis"}
-              >
-                {product.name}
-              </Text>
-              <Text
-                fontSize="14px"
-                color="#b8b8b8"
-                h={"30px"}
-                w={"198px"}
-                m={"6px 17px 6px 0"}
-              >
-                {product.description}
-              </Text>
-              <Text
-                fontWeight="bold"
-                fontSize="16px"
-                color="000000"
-                w={"108px"}
-                isTruncated
-              >
-                {currency(product.price)}
-              </Text>
-            </Box>
-            <Button
-              variant={"outline"}
-              borderColor={"black"}
-              borderTopLeftRadius={"8px"}
-              borderTopRightRadius={"0px"}
-              borderBottomRightRadius={"8px"}
-              onClick={onOpen}
-              borderBottomLeftRadius={"0px"}
-              alignSelf={"flex-end"}
-              fontSize={"12px"}
-              maxW={"90px"}
-              minW="90px"
-              h={"31px"}
-              m={"9px 0 0 17px"}
-              padding={"0"}
-              margin={"0 0 -1px -87px"}
-              color={"black"}
-            >
-              adicionar
-            </Button>
-            <Modal
-              blockScrollOnMount={true}
-              isOpen={isOpen}
-              onClose={onClose}
-              bg={"rgba(0, 0, 0, 0.5);"}
-            >
-              <ModalOverlay bg={"rgba(0, 0, 0, 0.5);"} />
-              <ModalContent maxW={"328px"} maxH={"216px"} borderRadius={"0"}>
-                <ModalHeader
-                  w={"296px"}
-                  h={"18px"}
-                  m={" 6px 0 0 16px"}
-                  fontSize={"16px"}
-                  letterSpacing={"-0.39"}
-                  textAlign={"center"}
-                >
-                  Selecione a quantidade desejada
-                </ModalHeader>
-                <ModalCloseButton />
-                <ModalBody>
-                  <Select
-                    w={"296px"}
-                    h={"56px"}
-                    m={"9px 16px 0"}
-                    p={"16px"}
-                    borderRadius={"4px"}
-                    border={"solid 1px #b8b8b8"}
-                    value={quantity}
-                    onChange={handleChange}
-                  >
-                    <option value="0">0</option>
-                    <option value="1">1</option>
-                    <option value="2">2</option>
-                    <option value="3">3</option>
-                    <option value="4">4</option>
-                    <option value="5">5</option>
-                  </Select>
-                </ModalBody>
-                <ModalFooter>
-                  <Text
-                    color="#4a90e2"
-                    onClick={onClose}
-                    textTransform={"uppercase"}
-                    w={"183px"}
-                    h={"19px"}
-                    margin={"7px 16px 16px"}
-                    fontSize={"16px"}
-                    letterSpacing={"0.39px"}
-                    textAlign={"right"}
-                    isTruncated
-                  >
-                    Adicionar ao carrinho
-                  </Text>
-                </ModalFooter>
-              </ModalContent>
-            </Modal>
-          </Flex>
-        </Center>
+        <ProductsCard
+          key={product.id}
+          photoUrl={product.photoUrl}
+          alt={product.name}
+          name={product.name}
+          description={product.description}
+          price={product.price}
+          onClick={() => onOpen(ModalAddButton)}
+        />
+
       );
     }
   });
@@ -210,128 +77,15 @@ export default function ResultPage() {
   const sideDishProductsMapped = products.map((product) => {
     if (product.category === "Acompanhamento") {
       return (
-        <Flex
-          border="solid 1px #b8b8b8"
-          borderRadius="8px"
-          margin={"0 0 -1px 43.5px"}
-          w="328px"
-          h="112px"
+        <ProductsCard
           key={product.id}
-        >
-          <Image
-            w="97px"
-            h="112.6px"
-            borderTopLeftRadius="8px"
-            borderBottomLeftRadius="8px"
-            src={product.photoUrl}
-            alt={product.name}
-          />
-          <Box ml="3">
-            <Text
-              fontWeight="bold"
-              fontSize="16px"
-              color="#e86e5a"
-              margin={"18px 16px 6px 0"}
-              maxW={"30ch"}
-              isTruncated
-              textOverflow={"clip"}
-            >
-              {product.name}
-            </Text>
-            <Text
-              fontSize="14px"
-              color="#b8b8b8"
-              h={"30px"}
-              w={"198px"}
-              m={"6px 17px 6px 0"}
-            >
-              {product.description}
-            </Text>
-            <Text
-              fontWeight="bold"
-              fontSize="16px"
-              color="000000"
-              w={"108px"}
-              isTruncated
-            >
-              {currency(product.price)}
-            </Text>
-          </Box>
-          <Button
-            variant={"outline"}
-            borderColor={"black"}
-            borderTopLeftRadius={"8px"}
-            borderTopRightRadius={"0px"}
-            borderBottomRightRadius={"8px"}
-            onClick={onOpen}
-            borderBottomLeftRadius={"0px"}
-            alignSelf={"flex-end"}
-            fontSize={"12px"}
-            maxW={"90px"}
-            minW="90px"
-            h={"31px"}
-            m={"9px 0 0 17px"}
-            padding={"0"}
-            margin={"0 0 -1px -87px"}
-            color={"black"}
-          >
-            adicionar
-          </Button>
-          <Modal
-            blockScrollOnMount={true}
-            isOpen={isOpen}
-            onClose={onClose}
-            bg={"rgba(0, 0, 0, 0.5);"}
-          >
-            <ModalOverlay bg={"rgba(0, 0, 0, 0.5);"} />
-            <ModalContent maxW={"328px"} maxH={"216px"} borderRadius={"0"}>
-              <ModalHeader
-                w={"296px"}
-                h={"18px"}
-                m={" 6px 0 0 16px"}
-                fontSize={"16px"}
-                letterSpacing={"-0.39"}
-                textAlign={"center"}
-              >
-                Selecione a quantidade desejada
-              </ModalHeader>
-              <ModalCloseButton />
-              <ModalBody>
-                <Select
-                  w={"296px"}
-                  h={"56px"}
-                  m={"9px 16px 0"}
-                  p={"16px"}
-                  borderRadius={"4px"}
-                  border={"solid 1px #b8b8b8"}
-                >
-                  <option value="0">0</option>
-                  <option value="1">1</option>
-                  <option value="2">2</option>
-                  <option value="3">3</option>
-                  <option value="4">4</option>
-                  <option value="5">5</option>
-                </Select>
-              </ModalBody>
-              <ModalFooter>
-                <Text
-                  color="#4a90e2"
-                  onClick={onClose}
-                  textTransform={"uppercase"}
-                  w={"183px"}
-                  h={"19px"}
-                  margin={"7px 16px 16px"}
-                  fontSize={"16px"}
-                  letterSpacing={"0.39px"}
-                  textAlign={"right"}
-                  isTruncated
-                >
-                  Adicionar ao carrinho
-                </Text>
-              </ModalFooter>
-            </ModalContent>
-          </Modal>
-        </Flex>
+          photoUrl={product.photoUrl}
+          alt={product.name}
+          name={product.name}
+          description={product.description}
+          price={product.price}
+          onClick={() => onOpen(ModalAddButton)}
+        />
       );
     }
   });
@@ -339,71 +93,16 @@ export default function ResultPage() {
   const drinkProductsMapped = products.map((product) => {
     if (product.category === "Bebida") {
       return (
-        <Flex
-          border="solid 1px #b8b8b8"
-          borderRadius="8px"
-          margin={"0 0 -1px 43.5px"}
-          maxW={"328px"}
-          maxH={"216px"}
-          key={product.id}
-        >
-          <Image
-            w="97px"
-            h="112.6px"
-            borderTopLeftRadius="8px"
-            borderBottomLeftRadius="8px"
-            src={product.photoUrl}
+        <Box>
+          <ProductsCard
+            key={product.id}
+            photoUrl={product.photoUrl}
             alt={product.name}
+            name={product.name}
+            description={product.description}
+            price={product.price}
+            onClick={() => onOpen(ModalAddButton)}
           />
-          <Box ml="3">
-            <Text
-              fontWeight="bold"
-              fontSize="16px"
-              color="#e86e5a"
-              margin={"18px 16px 6px 0"}
-              isTruncated
-            >
-              {product.name}
-            </Text>
-            <Text
-              fontSize="14px"
-              color="#b8b8b8"
-              h={"30px"}
-              w={"198px"}
-              m={"6px 17px 6px 0"}
-            >
-              {product.description}
-            </Text>
-            <Text
-              fontWeight="bold"
-              fontSize="16px"
-              color="000000"
-              w={"108px"}
-              isTruncated
-            >
-              {currency(product.price)}
-            </Text>
-          </Box>
-          <Button
-            variant={"outline"}
-            borderColor={"black"}
-            borderTopLeftRadius={"8px"}
-            borderTopRightRadius={"0px"}
-            borderBottomRightRadius={"8px"}
-            onClick={onOpen}
-            borderBottomLeftRadius={"0px"}
-            alignSelf={"flex-end"}
-            fontSize={"12px"}
-            maxW={"90px"}
-            minW="90px"
-            h={"31px"}
-            m={"9px 0 0 17px"}
-            padding={"0"}
-            margin={"0 0 -1px -87px"}
-            color={"black"}
-          >
-            adicionar
-          </Button>
           <Modal
             blockScrollOnMount={true}
             isOpen={isOpen}
@@ -458,7 +157,7 @@ export default function ResultPage() {
               </ModalFooter>
             </ModalContent>
           </Modal>
-        </Flex>
+        </Box>
       );
     }
   });
@@ -489,7 +188,6 @@ export default function ResultPage() {
         category={restaurantDetails.category}
         address={restaurantDetails.address}
       />
-
       <Text
         marginLeft="16px"
         alignSelf="center"
@@ -500,7 +198,6 @@ export default function ResultPage() {
       </Text>
       <Divider />
       {mainProductsMapped}
-
       <Text
         marginLeft="16px"
         alignSelf="center"
